Move generic UI methods to separate file
